54th Commit : Set Permission Guard Admin

+++ app/Http/Controllers/Admin/AdminAdminUserController.php
@@ -7,7 +7,6 @@ use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
 use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Mail;
 use App\Mail\AdminRegisterVerifyMail;
 use App\Http\Requests\Admin\AdminAdminUserStoreRequest;
@@ -15,6 +14,15 @@ use App\Http\Requests\Admin\AdminAdminUserUpdateRequest;
 
 class AdminAdminUserController extends Controller
 {
+    function __construct()
+    {
+        $this->middleware('permission:sytem admin user create,' . getGuardNameAdmin(), ['only' => ['create', 'store']]);
+        $this->middleware('permission:sytem admin user delete,' . getGuardNameAdmin(), ['only' => ['destroy']]);
+        $this->middleware('permission:sytem admin user index,' . getGuardNameAdmin(), ['only' => ['index', 'show', 'data']]);
+        $this->middleware('permission:sytem admin user restore,' . getGuardNameAdmin(), ['only' => ['edit', 'update']]);
+        $this->middleware('permission:sytem admin user update,' . getGuardNameAdmin(), ['only' => ['edit', 'update']]);
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -165,20 +173,29 @@ class AdminAdminUserController extends Controller
                     return '<div class="badge badge-danger">'  . __('No Action') . '</div>';
                 } else {
                     if ($query->status == 1) {
-                        return '
-                            <a href="' . route('admin.admin.edit', $query->id) . '" class="btn btn-primary btn-sm">
-                                <i class="fas fa-edit"></i>
-                            </a>
-                            <a href="' . route('admin.admin.destroy', $query->id) . '" class="btn btn-danger btn-sm delete_item">
-                                <i class="fas fa-trash-alt"></i>
-                            </a>
-                        ';
+                        if (canAccess(['sytem admin user update'])) {
+                            $update = '
+                                <a href="' . route('admin.admin.edit', $query->id) . '" class="btn btn-primary btn-sm">
+                                    <i class="fas fa-edit"></i>
+                                </a>
+                            ';
+                        }
+                        if (canAccess(['sytem admin user delete'])) {
+                            $delete = '
+                                <a href="' . route('admin.admin.destroy', $query->id) . '" class="btn btn-danger btn-sm delete_item">
+                                    <i class="fas fa-trash-alt"></i>
+                                </a>
+                            ';
+                        }
+                        return (!empty($update) ? $update : '') . (!empty($delete) ? $delete : '');
                     } else {
-                        return '
-                            <a href="' . route('admin.admin.restore', $query->id) . '" class="btn btn-warning btn-sm" data-toggle="tooltip" title="' . __('Restore to Active') . '">
-                                <i class="fas fa-undo"></i>
-                            </a>
-                        ';
+                        if (canAccess(['sytem admin user restore'])) {
+                            return '
+                                <a href="' . route('admin.admin.restore', $query->id) . '" class="btn btn-warning btn-sm" data-toggle="tooltip" title="' . __('Restore to Active') . '">
+                                    <i class="fas fa-undo"></i>
+                                </a>
+                            ';
+                        }
                     }
                 }
             })
