+++ app/Http/Controllers/Admin/AdminLanguageController.php
@@ -10,6 +10,15 @@ use App\Http\Requests\Admin\AdminLanguageUpdateRequest;
 
 class AdminLanguageController extends Controller
 {
+    function __construct()
+    {
+        $this->middleware('permission:language create,' . getGuardNameAdmin(), ['only' => ['create', 'store']]);
+        $this->middleware('permission:language delete,' . getGuardNameAdmin(), ['only' => ['destroy']]);
+        $this->middleware('permission:language index,' . getGuardNameAdmin(), ['only' => ['index', 'show', 'data']]);
+        $this->middleware('permission:language restore,' . getGuardNameAdmin(), ['only' => ['edit', 'update']]);
+        $this->middleware('permission:language update,' . getGuardNameAdmin(), ['only' => ['edit', 'update']]);
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -149,20 +158,29 @@ class AdminLanguageController extends Controller
                     return '<div class="badge badge-danger">'  . __('No Action') . '</div>';
                 } else {
                     if ($query->status == 1) {
-                        return '
-                            <a href="' . route('admin.language.edit', $query->id) . '" class="btn btn-primary btn-sm">
-                                <i class="fas fa-edit"></i>
-                            </a>
-                            <a href="' . route('admin.language.destroy', $query->id) . '" class="btn btn-danger btn-sm delete_item">
-                                <i class="fas fa-trash-alt"></i>
-                            </a>
-                        ';
+                        if (canAccess(['language update'])) {
+                            $update = '
+                                <a href="' . route('admin.language.edit', $query->id) . '" class="btn btn-primary btn-sm">
+                                    <i class="fas fa-edit"></i>
+                                </a>
+                            ';
+                        }
+                        if (canAccess(['language delete'])) {
+                            $delete = '
+                                <a href="' . route('admin.language.destroy', $query->id) . '" class="btn btn-danger btn-sm delete_item">
+                                    <i class="fas fa-trash-alt"></i>
+                                </a>
+                            ';
+                        }
+                        return (!empty($update) ? $update : '') . (!empty($delete) ? $delete : '');
                     } else {
-                        return '
-                            <a href="' . route('admin.language.restore', $query->id) . '" class="btn btn-warning btn-sm" data-toggle="tooltip" title="' . __('Restore to Active') . '">
-                                <i class="fas fa-undo"></i>
-                            </a>
-                        ';
+                        if (canAccess(['language restore'])) {
+                            return '
+                                <a href="' . route('admin.language.restore', $query->id) . '" class="btn btn-warning btn-sm" data-toggle="tooltip" title="' . __('Restore to Active') . '">
+                                    <i class="fas fa-undo"></i>
+                                </a>
+                            ';
+                        }
                     }
                 }
             })
